Added new redirect detection and handler
Referrer is now passed to background.js when a page is loaded

If the referrer is reddit, the new tab isn't actually on a reddit page, and it
doesn't already have the info already, it's a redirect.

If a redirect is detected, handleRedirect() is called, which gets the
chrome.history.VisitItem of the page and all the stored Info objects and
compares them.
If/when a match is found, the new (post-redirect URL) is added to the
existing Info object and the chrome.extension.onConnect handler is called again,
to determine if the bar should actually be shown.